Fix an error and a warning.
Use a separate variable to index through the hash digest, not the
variable used to index through the packet.

Make some variables holding one-byte quantities u_int rather than u_char
while we're at it - not much use in *requiring* them to fit into 8 bits
(the code might be better if we don't, and that eliminates a warning).