OpenFlow 1.0: enable decoding of nested frames
The packet that may be contained in an OFPT_PACKET_IN or OFPT_PACKET_OUT
OpenFlow message is a plain Ethernet frame. Make it possible to have
ether_print() process that frame, with a controlled verbosity level (see
a comment in print-openflow-1.0.c for details). Update an existing test
case to employ the new feature.