adopt MacOS deprecation workaround from FreeRADIUS
A MacOS build of tcpdump used to produce deprecation warnings on OpenSSL
function calls. Alan DeKok explained that this is the effect of Apple's
modifications to OpenSSL and that the very same problem has already been
solved in FreeRADIUS server source code with pre-processor macros.

Copy the macros into tcpdump header file and make use of them around the
functions that call OpenSSL functions.