Add -g3 on some platforms even if .devel isn't present.
Now that we're requiring autoconf 2.61 or later, AC_PROG_CC sets -g
along with -O2 for GCC-like compilers regardless of whether .devel is
present.  Add "include debugging symbols" options for MIPS C/DEC C
regardless of whether .devel is present; that's -g3, which is the
version that doesn't turn optimization off (we also turn on -O).

Also, don't change the ABI on IRIX depending on .devel.

I'm not sure why we were doing that; if somebody has a good reason to
continue doing it, please let us know what it is.

Expand some comments while we're at it.