fix flags printing tests
The test cases included the timestamp (due to missing -t) and failed
unless run in the same timezone as produced. The failures printed by
print-flags.sh did not make it into the final report because the script
always returned 0.

Strip the timestamps and replace print-flags.sh with some contents in
TESTLIST.