Qualify "length" when printing it.
In the "the TLV length is too short" message, we're printing the length
of the entire TLV; report it as "TLV length".  If we pass that test,
we've subtracted out the lengths of the T and the L, leaving only the
length of the V, so report it as "value length".