Don't recompile the filter if the new file has the same DLT.
Only recompile the filter - and only find a new printer and a new DLT
name - if, when reading multiple files with the -V flag, the new file
has a different DLT value from the old file.  With lots of small files,
that can make a significant performance difference.

Addresses GitHub issue #508.