Work around pcap_create() failures fetching time stamp type lists.
Normally, it succeeds whether the device actually exists or not, but, in
some versions (but not the latest version), it may fail on Linux trying
to fetch the list of time stamp types.  Check for that failure and, if
that's the reason for pcap_create() failing, return from
open_interface() so the caller can see if the interface name is actually
an interface index.