Clean up printing of Q.933 packets.
Check the packet length and the snapshot length as we parse the packet.
Extract each field as we go.

Support arbitrary call reference lengths, 0 to 15 octets.

Handle single-octet IEs correctly - don't look for a length.

Handle both locking and non-locking shifts correctly.  Don't assume that
the first octet after the message type is a shift and contains a
codeset.  We were doing that, meaning that we tended to think codeset 1
was being used (by misparsing an IE with a code of 0x51 as a shift to
codeset 1) when it wasn't - codeset 0 was being used; correctly handle
codeset 0.