Further IPv4-mapped IPv6 address cleanups.
Have a routine to check for those addresses, and use it in all cases.

Have a #define for the length of the fixed part of those addresses at
the beginning, and use it.

Fix a length check; Coverity noted that the check never failed, because
we were comparing the length of the prefix field in the packet with the
length of the IPv4 prefix, not the length of the IPv4 prefix plus the
fixed part at the beginning.