Clean up the crud supporting passed/failed count global variables as files.
Fetch the values from the files at the end of runComplexTests, rather
than at the beginning of runSimpleTests - they're already set to 0 if
you haven't run runComplexTests.

Set the variables to 0 right before writing their values to the files,
just to make it clearer that they're both being initialized to 0.

Read the files unconditionally in tests/crypto.sh, as we do tests
regardless of whether we were built with libcrypto or not.