Don't use -q, just send the standard output to /dev/null.
The default *grep in Solaris is *NOT* SUSv3-compliant, and doesn't
support -q; you have to use the one in /usr/xpg4/bin for that one.

Just use egrep (to make sure we support alternation, which we use),
don't pass it -q, and send its output to /dev/null, as we don't need to
see it.