Only print packets_captured as often as necessary.
Before this change tcpdump would, when run with -w and -v, print the
number of captured packets regardless if it had already printed the same
number, e.g. it could print "Got 0" again and again, triggering the
terminal emulator activity detection all the time and making it useless.
Only print the string when it is different from the previous one.