Use nd_ types, clean up ICMPv6 dissection.
Use nd_ types for various packet structures, and use EXTRACT_ macros as
necessary.  Get rid of & in EXTRACT_ macros if not necessary.

The 1-byte pad RPL suboption is called PAD1, not PAD0, in the RFC.

Rename rpl_dio_printopt() to rpl_printopts() because 1) it's for all RPL
messages and 2) it prints multiple options.  Clean up its processing
loop.

Pass rpl_print() the ICMPv6 code, not a pointer to the full header; all
it needs is the code.

Use %u to print unsigned values.

In Node Information node address lists, the TTL comes before the
address, as per the RFC.  Dissect it that way.