Make 1-element arrays for fields that may repeat.
For various opaque LSAs, not only is the value of the TLV a "may
repeat", the TLVs *themselves* may repeat.

Also, pass a pointer to the TLV to ospf_print_grace_lsa() and
ospf_print_te_lsa(), rather than a pointer to the type field, as they
dissect a sequence of TLVs.

Hopefully, that will address Coverity CID 1426916, 1426917, and 1426920.