Fix some things.
decode_labeled_vpn_l2() should return the full length of the item,
including the header; fix it to do so.

Fix calculation of length-in-bytes from length-in-bits.

Add additional length checks.

Do checks of unsigned values against zero as "not equal to zero" rather
than "greater than zero" (they should compile to the same code as, for
unsigned values. they mean the same thing, but that makes it a bit
clearer than it's not as if we're testing for negative values).