Fix check for rpc/rpcent.h.
Autotools' check in AC_CHECK_HEADERS() apparently either doesn't
actually try to include it or somehow arranges to define stuff that it
requires on FreeBSD.  CMake's check in check_include_file() just tries
to include it, which fails on FreeBSD because __BEGIN_DECLS isn't
defined.

The man page for getrpcbynumber() says we just need to include
<rpc/rpc.h>; perhaps there are platforms on which we require both, and
that's what we do, so test it with rpc/rpc.h if we have rpc/rpc.h.