Squelch more narrowing warnings.
Add an ND_BYTES_BETWEEN() macro that computes how many bytes are
present, starting at the second argument and running up to (but not
including) the first argument, and returns that as a u_int (cutting it
to 32 bits on LP64 and LLP64 platforms).

Use that, including using it as a replacement for SMB's PTR_DIFF().