Clean up the switch-tagged header code.
Use MAC_ADDR_LEN in expressions representing offsets and lengths.

Don't declare the Ethernet header structure ourselves in the Broadcom
code, just print it the same way it's printed in the Marvell code.  Make
the two routines similar in other places as well.

Use "Unknown" rather than "unregistered" when printing the Ethernet type
field in the EDSA header - the value may well be registered even if it's
not in ethertype_values, it's just not a value we happen to have in that
table.