Clean up "make check".
No need to make a "tests" directory under the build directory any more,
so don't do so.

As we're not cd'ing in the command line run from the Makefile (just in a
command run for command substitution), there's no need to put it in a
parenthesized subshell.

Again, as we're not cd'ing to the tests directory, if SRCDIR isn't set,
assume the source directory is the current directory.

In the shell scripts we're running, look for TESTonce, the input files,
the output files, and the -E input files from ${srcdir}/tests.  Make
sure we set srcdir in those shell scripts.