RSVP: Add a test capture file
This capture file triggered a buffer over-read with the 4.9.2 release
fixed in the 4.9.3 release.

With the new GET_ macros, the problem no longer exists.
Thus no fix needed.

The pcap file was updated to pass the sanity checks and to avoid
pcap_loop() error messages.

Update the output of the test accordingly.

Partial update from bea2686c296b79609060a104cc139810785b0739
in 4.9 branch.