OSPFv3: Add a test capture file
This capture file triggered a buffer over-read with the 4.9.2 release
fixed in the 4.9.3 release.

With a bounds check previously added, the problem no longer exists.
Thus no fix needed.

The pcap file was updated to avoid pcap_loop() error messages.
It was rename from pcapng to pcap because it is.

Update the output of the test accordingly.

Partial update from e01c9bf76740802025c9328901b55ee4a0c49ed6
in 4.9 branch.