Move some code to better show how to handle pcap_activate() failure.
If you call pcap_activate() and it fails, you still have a pcap_t from
pcap_create(), and should close it.

Moving the code after the error() calls doesn't make a difference, as
error() exits, but it may make it clearer to those reading tcpdump.c for
help in figuring out how to use libpcap that you should close the pcap_t
if pcap_activate() fails.

(In the future, there may also be the option of changing some options
and trying again, e.g. changing the user name or password for a remote
capture.)