Fix check of authentication algorithm.
If you split a string by overwriting the separator character with a NUL,
and you want to compare the first part with various values, use a
pointer to the beginning of the string, not to the NUL you just dropped
in.

This is a bug found while testing some cleanups for GitHub issue #814;
it's necessary for the cleanups to work.