Do some tests based on the type of floating-point arithmetic tcpdump does.
Add a --fp-type flag to tcpdump, which causes it to do a floating-point
operation and, based on the result of the operation, prints out
"FPTYPE{n}", where {n} is a number indicating the result.

Have tests/TESTrun run "./tcpdump --fp-type" and set a HAVE_ key based
on that.  Run some tests only for FPTYPE1.