ether: clean up MACsec processing.
Print the length early if we're printing the link-layer header.

If the payload is encrypted or otherwise modified, print it out as raw
data.

If the payload is not encrypted or otherwise modified, and we didn't
have a problem printing the header, fetch the type/length field
following the MACsec header, skip past it, and continue, rather than
looping back - there shouldn't be multiple MACsec headers, as far as I
know.  (If that's not the case, go back to looping.)