TESTrun: fix some problems when run on Windows.
Run tcpdump, not windump - we build it as tcpdump, not windump, and
there's not really a reason to have a separate windump program.

For now, at least, assume a Visual Studio debug build, so it's in the
Debug subdirectory.

Use that path for *all* calls to tcpdump.

Use "type" rather than "cat" to print files on Windows.

It still needs to be changed to, for example, not use sed.