TESTrun: process the system return value differently on Windows and UN*X.
We can use the POSIX functions on UN*X, but not on Windows.  Use them on
UN*X, but extract the exit status manually on Windows.

Put in a pile of comments about the return status of commands, command
interpreters, and the Perl system function on UN*X and Windows.

If system simply fails to run at atll (returns -1), put a message into
the faked output file.

Get rid of some duplicate initializations while we're at it.