OpenFlow 1.0: Fixup some indentation. [skip ci]
In this file an ASCII tab character means "one level of indentation,
whatever amount of whitespace an editor uses to _display_ it", and a
sequence of several spaces means "exactly this many spaces". Sometimes
they are mixed together for a specific purpose, which is easy to see in
conditional complex print statements: the ND_PRINT() line is one more
level of indentation deeper than the if(), and the wrapped remainder in
addition to that is as many spaces deeper as necessary to align with the
first argument. It is not the simplest thing to explain, but easy to see
how it works in the actual code.

Wrap and unwrap some lines to accommodate the code churn, assuming that
one level of indentation is worth at most 8 positions and that most lines
should look at most 80 characters long unless there is a good reason to
keep them longer.