Fix a few bitmap printing bugs.
bittok2str_internal() used to match on a single walking 1, so a bitmap
token that had more than one bit set would never "match all" or "match
any". Replace the bit-walking loop with a single bitwise test, that's
faster and it provides "match all" for multiple-bit tokens if required.
Also a 0 token would always match a 0 value, which makes no sense in
bitmaps, so fix a 0 token to never match. Update the comment and fixup
some indentation.

This change had immediately chased a bug out of print-udld.c, where the
tokens were wrong in udld_flags_values[]. Fix that, list the MBZ bits and
rename the array for clarity. Update output for two tests.

In the course of this debugging it became clear that in print-aoe.c I got
both the tokens and the labels wrong in aoev1_aflag_str[]. Fix that,
rename the listed MBZ bits for brevity and rename the array for clarity.
Update output for one test.