EAP: add some more length checks.
Check to make sure we don't go past the length value in the header.

Note that the string in an Identity message is optional (so we just
don't bother printing it if it's zero-length), but the string in a
notification message isn't (so report an error if it's zero-length).

(cherry picked from commit 50dfa7ecebfef0f954f3ac2e6bbbf84ab468f848)