Use "git archive" for the "make releasetar" process
Use the release tag if it exists or use HEAD.

Remove the "autoreconf -f" command, because it changes the configure file
locally in an uncontrolled way (runstatedir, #define LARGE_OFF_T, etc.)
depending on the autoconf version. This command is run in the release
process before a commit and we can choose the parts that will be added
in the commit.

Note:
The following target (rcX) must be updated ou removed in a next step
because it was using "autoreconf -f" (via "make releasetar") and it
changes the VERSION and configure files locally.

This change
1) Ensures that we only release files from tag/HEAD, not locally
   modified ones.
2) Avoids disclosing personal data such as the username/group of the
   local user.
3) Puts by default a umask of 0002, which turns off the world write bit
   on files in the archive.
4) Avoids problems on some OSes (no more tar, Git builtin tar.gz
   handling).

(cherry picked from commit c378c2a04cdb3c8ec35341ca503cf1a2889a77d5)