CI: Introduce and use TCPDUMP_CMAKE_TAINTED
The new environment variable provides means to declare all or some tcpdump
cmake run of a specific environment (CI or a working copy) as tainted.
Add explicit conditions to exempt the cmake run that emit warnings
now and request the compiler to treat warnings as errors *iff* the cmake
run is not tainted. This way if a cmake run that is warning-free now
degrades later, CI will fail it.

The treat warnings as errors cmake option used is:
-Werror=dev
       Make developer warnings errors.

       Make  warnings  that  are  meant  for  the  author of the CMake-
       Lists.txt files errors. By default this will also turn on depre-
       cated warnings as errors.

(cherry picked from commit 625657ada89f6ab9056e3f918edb7052d3067438)