configure: check for gethostbyaddr(), not gethostbyname().
We use gethostbyaddr() to translate IP addresses in packets to names; we
don't use gethostbyname() to translate names to IP addresses (libpcap
does the latter when compiling filters, so we leave that up to it, and
nothing's left for us to do).

This probably has the same effect as checking for gethostbyname(), but
it makes it a bit clearer why we'd check for it.

(cherry picked from commit b1ccc20ccc9bdd73ed065db61ab9cb5b74e5acb4)