ptp: Parse major and minor version correctly
In IEEE1588-2008, the lower 4 bits of the field were defined as the PTP
version, and the upper 4 bits were marked as reserved. As of the 2019
version the upper 4 bits are defined as PTP minor version. Previously,
the whole byte was assumed to be the version by Tcpdump, resulting in
PTP v2.1 being parsed as v18 and not printing the rest of the
information.

With this commit the version is parsed correctly and the packet is
displayed as v2. Note: I do not have access to the 2019 version, so I
unfortunately can't add whatever new features it adds.

Signed-off-by: Casper Andersson <casper.casan@gmail.com>

(cherry picked from commit 63134a45c19b49ea113f2c88fa7367be13997ac0)