RADIUS: Rename a test with an invalid length
The RADIUS length is invalid. The UDP length is 64, thus the UDP
payload is 56 bytes, but the RADIUS length is 57.

tcpdump doesn't currently show it because of the following lines in
radius_print():
   if (len > length)
          len = length;

This code will be fixed in a future update.

(cherry picked from commit f64a4a5f49bcbcb996820b566d7ffe9f7cefe4f3)