Autoconf, CMake: Get the size of a void * and a time_t
With this change we know:
whether this is a 32-bit or 64-bit build ;
whether the time_t size is 32-bit or 64-bit.

At least with CMake, the SIZEOF values could be 0, if somebody's
doing a fat build on macOS and that includes both 32-bit and 64-bit
instruction sets.

(cherry picked from commit 170cbc8c6d484387ec7ae19ef9f4bd14dd0845e7)